start moving things into pages
