--- src/app/api/apps/route.ts
@@ -1,9 +1,0 @@
-import { NextRequest, NextResponse } from "next/server";
-
-import { upsertApp } from "@/app/app";
-
-export async function POST(req: NextRequest) {
-  const app = await req.json();
-  await upsertApp(app);
-  return new NextResponse();
-}
