+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import { Inter, Roboto_Mono } from "next/font/google";
 import { Toaster } from "@/components/ui/sonner";
 
-import "./globals.css";
+import "@/pages/globals.css";
 import { Footer } from "@/components/Footer";
 import { Metadata } from "next";
