+++ src/pages/apps/[id].tsx
@@ -1,0 +1,5 @@
+import Layout from "@/components/Layout";
+
+export default function Page() {
+  return <Layout>Hello World! this is an app</Layout>;
+}
